- Css files are not put in another folder, they will stay with the same folder as the component they apply to. - not much to do as we have done this before  , I moved the About us component in the component folder. - removed some unused yml files in the workflow folder
