--- .github/workflows/AuthentificationTest.js.yml
@@ -1,34 +1,0 @@
-# This workflow will do a clean installation of node dependencies, cache/restore them, build the source code and run tests across different versions of node
-# For more information see: https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-nodejs
-
-name: AuthentificationTest
-
-
-on:
-  push:
-    branches: [ "profile-admin-isa" ]
-  pull_request:
-    branches: [ "profile-admin-isa" ]
-
-jobs:
-  auth_check:
-    runs-on: ubuntu-latest
-
-    steps:
-    - name: Check Firebase Authentication Information
-      id: auth_check
-      run: |
-        # Pattern to match Firebase configuration objects
-        pattern='apiKey:[^,]+,.*authDomain:[^,]+,.*projectId:[^,]+,.*storageBucket:[^,]+,.*messagingSenderId:[^,]+,.*appId:[^,]+,.*measurementId:[^,]+'
-
-        # Check if the pattern exists in any file in the pull request
-        if ! git grep -E -q "$pattern"; then
-          echo "Firebase authentication information is missing from the pull request."
-          exit 1
-        fi
-
-      shell: bash
-
-    - name: Set GitHub Action Status
-      if: ${{ failure() }}
-      run: echo "::error::Firebase authentication information is missing. Please add it to a secret file."
