+++ .github/workflows/main.yml
@@ -2,7 +2,7 @@
 # For more information see: https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-nodejs
 
 #INDENTATION MATTERS
-name: IntegrationTest_1
+name: main_intergation_Pipeline
 
 # this test will run on every push or pull request made to main branch
 # note : the test will only run on branches where the .yml file IS !!
