+++ src/Pages/AboutUs.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import Card from '../components/Card';
-import About_component from '../About_component';
+import About_component from '../components/homepage/About_component';
 function AboutUs() {
   return (
     <div>
